generic: 6.6: add pause configuration for pcs-mtk-lynxi
When the ethernet mac is configured managed = "in-band-status", the
phy-handle will be bound to the internal pcs phy, and the pause
configuration will be ignored. This will make some Mediatek Filogic
Devices with SFP can't enable flow control by default. This patch adds
the pause configuration by default. The user can still disable the
flow control configuration using ethtool if needed.

Tested on: CMCC RAX3000M with a hack PCB that connects USB3.0 to SFP

Signed-off-by: Yangyu Chen <cyy@cyyself.name>